Start writing the API doc

+++ src/main/java/com/qwazr/search/index/SettingsDefinition.java
@@ -1,12 +1,12 @@
 /**
  * Copyright 2015 Emmanuel Keller / QWAZR
- * <p>
+ * <p/>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * <p>
+ * <p/>
  * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ * <p/>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -23,7 +23,7 @@ public class SettingsDefinition {
 
 	final public Integer max_simultaneous_write;
 	final public Integer max_simultaneous_read;
-	final public Integer max_size;
+	final public Long max_size;
 	final public FileClassCompilerLoader.JavacDefinition javac;
 
 	public SettingsDefinition() {
